Limit emails to a small group of engineers on firmware test failures
Running FAFT (firmware tests) on lab is not ready yet. We found the
firmware is not re-flashed before running the tests. It causes a lot
of tests failed and keeps sending annoying emails to the whole firmware
group. We should limit the emails to a small group first. When the
tests run stably, revert it back to the whole firmware group.

BUG=chromium-os:22475
TEST=none

Change-Id: Id22f5f16e692e0e94b9032184ea6fb12fa2b193a
Reviewed-on: https://gerrit.chromium.org/gerrit/44384
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>