autotest: Add mechanism to block on xmlrpc_connect completion
As mentioned in the xmlrpc_connect documentation, the process of
starting an xmlrpc server does not imply that the server is ready
to serve requests.  Provide an optional mechanism to ensure that
the server is ready to service requests before xmlrpc_connect returns.

TEST=Note that calling xmlrpc_connect with the optional mechanism
enabled allows the caller to remove arbitrary time.sleep calls.
BUG=None

Change-Id: Iccdb0721e58397b86d54f34c695c9729313bc714
Reviewed-on: https://gerrit.chromium.org/gerrit/44417
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>