Autotest add timeout to RPM client and server side code.
It appears that calls to the RPM infrastructure can get hung and
will get stuck as the calls are blocking. This will cause hosts to
be held by jobs forever and we can not release them. In order to
get around this I added timeout code to the client that makes the
calls and the server side code that processes the calls.

BUG=chromium-os:39043
TEST=ran the infrastructure locally with a low timeout value and
ensured both the client and the server side caught the timeout so
that the client side code could continue and the server side code
would no longer process that request.

Change-Id: I6bf5de959336cda79a0939d00348db11087ec8cc
Reviewed-on: https://gerrit.chromium.org/gerrit/44535
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>