Revert "[autotest] Schedule tests against lumpy-pgo-release builds."
Although this is the correct way to parse, it causes us to schedule
tests against lumpy-pgo-release, which is not what is wanted. There
needs to be a large amount of other work to fix this handling of
lumpy-pgo-release properly, so in the meantime, the old form of breakage
(scheduling against lumpy-release when we see lumpy-pgo-release) is
actually preferable.

This reverts commit c3d4b7950036abb6c89f884d529c8d8839e23eed

Change-Id: Id0d352c56560f6baccf719dc766072cee904d6e6
Reviewed-on: https://gerrit.chromium.org/gerrit/44544
Commit-Queue: Alex Miller <milleral@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>