Increase debugging output in case of power_suspend hwclock problem
We are currently seeing a rare flakyness of suspend tests on Parrot and
Butterfly: the 'hwclock --utc --debug' check during initialization of
the power_suspend.py library gives unexpected output, indicating that
the /dev/rtc device node was not found or acted incorrectly.

This patch adds some debugging output to that failure path, in the hope
that this will give us more information about what is going on.

BUG=chromium-os:39486
TEST=None

Change-Id: I3f2b3f2674c5eb50deaccc1e119c96ea3228c9b2
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/44589
Reviewed-by: Derek Basehore <dbasehore@chromium.org>