Increase character limit in autotest DB for perf keys from 30 to 256 characters.
This change is accomplished by adding a new migration file to the set of
autotest DB migrations (with this change, we're now at version 72).

This change is required for new Telemetry perf tests -- written by the chrome
team -- that we'll soon run on chromeOS.  Some of these tests have perf keys
that are considerably longer than 30 characters, but are less than 256
characters.

Also removed an unused command-line option ("-a") from migrate.py,
the script that is used to perform the database migrations.

Also added a bunch of docstrings and removed some unused imports, to
satisfy pylint.

BUG=chromium-os:39494
TEST="python migrate.py sync", running on my local machine.  Verified that
I can migrate up and down from version 71 to 72, and when doing so, the
column "tko_iteration_result.attribute" changes from varchar(30) to
varchar(256) as expected.  Similarly, this change is reflected in the
views "tko_perf_view.iteration_key" and "tko_perf_view_2.iteration_key".

Change-Id: I4fa7524f6d732a5ff8f4950445d64d3f20a8c630
Reviewed-on: https://gerrit.chromium.org/gerrit/44647
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>