Increase max length of TKO model for tko_iteration_result.attribute to 256.
In https://gerrit.chromium.org/gerrit/#/c/44647/ , the max length of
a perf key in the autotest database is increased to 256 characters.
The current CL makes the corresponding change to the max length in the
corresponding TKO model.

Also added docstrings and removed one import that appears to be unused,
to satisfy pylint.

BUG=chromium-os:39494
TEST=None

Change-Id: I54d1089f225c0a92f0477fe4cc8508d008aa7977
Reviewed-on: https://gerrit.chromium.org/gerrit/44668
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>