[autotest] add retries field to create job view
A job created in the afe's create job tab can now have its test_retry
value specified. If the field in the afe is left blank, the number of
retries used is taken from the RETRIES value that was parsed from the
control file at test-import time (or the maximum of this value ot
over all the selected jobs, if multiple jobs are selected).

BUG=chromium-os:37158
TEST=Ran './utils/compile_gwt_clients.py -c autotest.AfeClient' to
recompile frontend. Then, 'sudo /etc/init.d/apache2 restart'. Verified
that the above behavior holds, and that the clone job button populates the
retry field with the value from the job to be cloned.

Change-Id: I6f28a0d9e1eb8a39fae0d1669bef846c65c99198
Reviewed-on: https://gerrit.chromium.org/gerrit/45164
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>