Make idle suspend stress subtests wait longer to synchronize each other
power_SuspendStress/control.idle* consists of two subtests that wait for
each other to start at the same time. Unfortunately, this hardly ever
works in current lab runs because our Chrome automation framework is so
horribly broken that it takes 40+ seconds to login.

This patch increases the duration these tests wait for each other's
initialization to work around that. It also activates the suite for the
per-build test (because I actually just forgot that last time).

BUG=chromium-os:38063
TEST=Run it.

Change-Id: Ib58c8628d8a9cd9d8462b559dd7d2532b84e6b0c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/44783
Reviewed-by: Mark Koudritsky <kamrik@chromium.org>