Abstract out operations needed for Servo recovery.
Create a new class that provides a board-independent interface to
basic "power off" and "power on" operations used during Servo
recovery.  The new class is meant to allow autotest repair to work
successfully regardless of the OS software is working on the DUT,
and without requiring callers to know board details.

This change is preparatory only; it doesn't affect the list of
supported boards.  Before and after this change, Servo repair is
restricted to just 'x86-alex' and 'lumpy'.  Additional changes will
follow to add support for additional boards.

BUG=chromium-os:35804
TEST=platform_InstallTestImage on a locally attached Alex
TEST=manually trigger repair for an Alex in a local autotest installation

Change-Id: I05a92726d7c3c21955e7a42cf16142c9edac066d
Reviewed-on: https://gerrit.chromium.org/gerrit/45159
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>