[autotest] smoke suite control file should be server side
This CL is mostly cosmetic, since the TEST_TYPE for a suite is never
actually used (suite control files are not imported through test
importer, which parses control file variables, nor are the control files
parsed before kicking off a suite with run_suite or
run_remote_tests.sh). All the other suites have this set as server side
though, and the control file is indeed being exec'd on the server, so
it's makes more sense to label this as a server side test.

BUG=None
TEST=ran a smoke suite on local autotest instance, kicked off with
run_suite.py

Change-Id: I145daf3bb759a40acb89449cf13c93c1c5d6793e
Reviewed-on: https://gerrit.chromium.org/gerrit/45390
Tested-by: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>