Modify our support of --image to work with our devservers.
Add a pathway to the --image code in site_host to stage anything that
is passed to it via --image that does not start with http://.

Added code to properly update job_repo_url in site_host rather
than in the try-jobs we spawn off. This makes it so that repaired
images as well as installs via --install have the proper job_repo_url
if they exist in the AFE.

Updated AfeClient.html to show examples of what can be specified in
Image Url.

Update old way of setting REPO URL in site_autotest.py to only be
used if someone is specifying a HTTP URL themselves, otherwise use the
AFE.

TEST=ran autoserv:
 ./autoserv --image=stumpy-release/R27-3836.0.0 -m scottz-stumpy1 -s
 tests/sleeptest/control
and
./autoserv
--image=http://172.22.50.205:8082/update/stumpy-paladin/R27-3837.0.0-rc2
 -m scottz-stumpy1 -s tests/sleeptest/control
and with a bad image which fails after a short while with an error about
the build path.
./autoserv --image=x86-mario-release/R27-3834.dd0.0 -m scottz-mario1 -s
 tests/sleeptest/control
Test on a non-afe host:
./autoserv --image=x86-mario-release/R27-3834.0.0 -m 172.29.151.133 -s
 tests/sleeptest/control
Tested repair update as well just to make sure I didn't break anything.
./autoserv -R -m scottz-mario1 # with a modification to make verify
fail always.
Ran tests through the frontend.
BUG=chromium-os:38653

Change-Id: I4b620a4332c695f4c972a8f8c112b4bf1dd6ba9a
Reviewed-on: https://gerrit.chromium.org/gerrit/45419
Tested-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Scott Zawalski <scottz@chromium.org>