[autotest] allow 0 test retries in create_job view
This CL fixes a parsing bug which was preventing jobs from being created
in the autotest afe create_job when the value 0 was specified for
test_retries.

Also, Java is verbose.

BUG=chromium:196247
TEST=compile frontend with utils/compile_gwt_clients.py. Then restart
apache. Then verify that able to create jobs with test_retry field set
to zero, and that leading or trailing spaces are ignored.

Change-Id: I46e4fc1d0e6cef8581a2941a6f9a343e4335aa6d
Reviewed-on: https://gerrit.chromium.org/gerrit/45439
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>