Switch utils.get_board() from HWIDs to /etc/lsb-release
This patch changes utils.get_board() to retrieve the board name from the
/etc/lsb-release file, instead of reading the HWID via crossystem. HWIDs
have proven unreliable on some old Mario boards, and are in general more
complicated to parse. As Autotests are mostly run in the lab, we can
expect machines to be imaged correctly and contain the right board name
in their lsb-release.

BUG=chromium:188841
TEST=Run power_Resume repeatedly on a Lumpy. If you see it suspend more
than once for a single test run but still return a GOOD result in the
end, you know that this change works.

Change-Id: Ib76413807e207ccdc7603b1fd7b8e0d208627350
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/45627
Commit-Queue: David James <davidjames@chromium.org>