Support wildcards in list of perf keys to plot on chromeOS perf graphs.
When generating chromeOS perf graphs, a JSON file specifies which perf keys
from which tests to plot on which graphs.  That JSON file now supports the
use of wildcards, so you don't need to specify exact perf keys in the JSON
file for them to show up in perf graphs:

1) *metric: Matches perf keys that end with "metric".
2) metric*: Matches perf keys that start with "metric".
3) *metric*: Matches perf keys that contain "metric".
4) metric: Matches only the perf key "metric" (exact match).
5) *: Matches any perf key.

BUG=chromium:188811,chromium:196265
TEST=Verified that when generating perf graphs from the perf results in
the past 7 days (which includes certain perf keys that are now longer than
they used to be), the newly-longer perf key results are now properly
showing up in the graphs with the use of wildcards, whereas before this change,
they were not showing up properly.

Change-Id: Id1c233f8db81f942f2fd91c7a2dc38bebedbfdb3
Reviewed-on: https://gerrit.chromium.org/gerrit/45507
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>