firmware_TouchpadMTB: make drumroll gesture file name backward compatible
We had removed the "normal" and "slow" variations on drumroll gesture
in a previous patch, while keeping the "fast" one. In the current
config file, it specifies "variations=None", the gesture file name
would look like

  drumroll-XXX.dat

which is not compatible with the already existing gesture files
which looks like:

  drumroll.fast-XXX.dat

In order to be able to replay the existing gesture event files,
we would like to set "variations=(GV.FAST,)," so that the file
name is backward compatible.

BUG=chromium-os:35024
TEST=Conduct the firmware tests on a chromebook machine. Observe
that it only tests the fast drum-roll gesture variation. The normal
and slow drum-roll ones have been removed. Note that the gesture
file name looks like: "drumroll.fast-XXX.dat".

Change-Id: Iab7a5599888d397f4209acfcbcc25759e6c24cf3
Reviewed-on: https://gerrit.chromium.org/gerrit/45576
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>