Add a touchscreen device option for FW test
To let the touch FW test run for touchscreens as well as touch pads we
need a flag to toggle between them.  This patch adds --touchscreen to
the OPTIONS for the test suite.  If it's set it reconfigures the Touch
device code to look for touchscreens instead of touchpads.

Obviously some wording in the code and display should probably be
changed since they all refer only to touchPADs but this is a bare-bones
working fix to get started heading in the right direction.

BUG=chromium:222247
TEST=I ran the FW test with and without the --touchscreen flag on a Link
and it worked both ways.
CQ-DEPEND=I3c70c7c3c40b8dddad3362207abe0848a41c1f4b

Change-Id: I4f9f3c4427d03ddb1f92ef1991d8362d9914f34e
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/45916
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>