autotest: Disable desktop_VideoSanity in BVT
This test fails with a kernel crash on Stout, which in turn causes
Autotest to fail the Canary build rather than ignoring the result
because of the experimental label.  There is a already a tracking high
priority bug to fix this error, and the TPMs are fine with releasing
canaries with this bug on Stout.

TEST=None
BUG=chrome-os-partner:18231

Change-Id: If94b6fb69316143ceb321414b87c45e2a9789ab5
Reviewed-on: https://gerrit.chromium.org/gerrit/46138
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>