Deal with empty returns correctly in test_image.
Fixes a bug in:
./full_release_test.py --specific 3833.0.0 --npo 3881.4.0 --all_boards

Where is a payload is missing for one board, it breaks the run for all
boards (in this case stumpy-release). This is because with my recent changes
I failed to honor the None or [] depending on single = True | False given in
the docstring.

BUG=None
TEST=Ran it with above -- gave warning about missing stumpy payload but
continued.

Change-Id: I4ce6ca30a2533622d35cb35a49fbe6f33bc4687c
Reviewed-on: https://gerrit.chromium.org/gerrit/46144
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>