Autotest: Make sure RPM-related tests ensure the outlet is ON.
Seems that tests that turn off outlets may fail without restoring them.

To address this any test that interacts with the RPM_Infrastructure
will add a host_attribute and in cleanup if it sees that attribute will
call power_on() to ensure it is turned on.

BUG=chromium:220670
TEST=Ran an RPM related test, ensured cleanup called power_on() followed
it with a sleep test and ensure cleanup did not call power_on().

Change-Id: Ia6f16e2b27317aa537be0737fd06f8e270709698
Reviewed-on: https://gerrit.chromium.org/gerrit/46158
Commit-Queue: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>