restart update-engine -> stop/start update-engine.
Some DUTs that do not have update-engine running properly initially
are failing repair due to the call to restart update-engine. This has
been changed to "stop upodate-engine; start update-engine" to avoid
this.

At this point if we end up in repair failed we should assume that
a repair is required as our verify is not all inclusive.

TEST=ran on local DUT
BUG=chromium-os:223081

Change-Id: I18c0b3703c6bf34df5aaaeff7980deda26dd051b
Reviewed-on: https://gerrit.chromium.org/gerrit/46265
Tested-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>