New baseline format for test, + latest baseline updates
This also tightens the list of what we look for, on the nss side,
to be limited to things marked as trusted for SSL specifically.
This accounts for nearly all of the previous "nss-only" items in
the test baseline. A bugfix related to cases where nsscertutil
returned more than 1 cert also closes someof the incorrectly-reported
gap.

Also lint fixes.

BUG=chromium:221667
TEST=security_RootCA passes on ToT/lumpy

Change-Id: I37d3d5cd7236017325c6e8cf109f37f056307dbe
Reviewed-on: https://gerrit.chromium.org/gerrit/46336
Tested-by: Jim Hebert <jimhebert@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Jim Hebert <jimhebert@chromium.org>