Temporarily move desktopui_VideoDecodeAcceleration to video suite.
This particular test is failing on devices in a way that causes the DUT
to reboot or no longer be reachable. Autotest currently treats this as
a hard failure regardless of this test being experimental or not.

By moving this to video we will insulate the tree from failure until
either this test is fixed or Autotest's behavior can be modified at this
level to avoid calling any failure from an experimental test fatal.

TEST=site_utils/suite_enumerator.py -a . video to confirm the suite showed up
there.
site_utils/suite_scheduler/suite_scheduler.py --sanity
BUG=chromium:188217

Change-Id: If3c8626736efffdcd36f31df7c422e4a78cd804e
Reviewed-on: https://gerrit.chromium.org/gerrit/46415
Tested-by: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Pawel Osciak <posciak@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Scott Zawalski <scottz@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>