[Autotest] Unexpeted reboots should not lead to unexpected ABORTS.
If the test failed inconclusively we should try and diagnose the
DUT. If we successfully diagnose it to either be un-pingable,
un-sshable or to have rebooted unexpectedly we should fail. Only
if we are unable to diagnose it (i.e everything seems normal but
the test still failed) do we end with an unexpected abort.

TEST=Induced crashes in power resume and made sure we diagnose
     them appropriately. Also made sure we END FAIL in these situations.
BUG=chromium:188217

Change-Id: I0ff60248abe0f1d0b2b0c24a268748b51c4318ae
Reviewed-on: https://gerrit.chromium.org/gerrit/46608
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>