Don't break nmo testing when we know the target branch and the branch updates.
This is part of getting nmo testing on the buildbot. This test can't break
every time we branch. The limiting factor here is that we use release_config
to determine the branch of the src image in a delta payload. We fix this by
parsing the delta for the branch.

BUG=chromium:223783
TEST=Remove R26 from release config and used archive_url + npo/nmo and it
worked.

Change-Id: I15ce652be052b64adadaffa11dcd1b27b70b1aca
Reviewed-on: https://gerrit.chromium.org/gerrit/46462
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>