power_DarkResumeShutdownServer:
This makes sure that the system is resumed and that the AC power is on for the
machine in finalize. This makes sure that if we run into a failure when the
power is off, we still have the power on when the test finishes.

BUG=chromium:220352
TEST=Run test with failure introduced after we turn the AC off

Change-Id: Ie7b748b7b57c8d7f198c2e2bd2a1b735f3f36c3c
Reviewed-on: https://gerrit.chromium.org/gerrit/46898
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>