firmware_TouchpadMTB: upload gesture data to google storage server
In this patch, we download and install gsutil package if not yet
onto a chromebook machine. After the firmware tests have been
finished, the gesture data files are uploaded to the bucket
chromeos-touchpad-public on the google storage server. This bucket
is for public use and a private bucket will be created later for
internal use.

BUG=chromium:224228
TEST=perform the firmware tests.
$ cd /usr/local/autotest/tests/firmware_TouchpadMTB
$ DISPLAY=:0 python main.py

After the tests have been finished, check
"https://storage.cloud.google.com/?arg=chromeos-touchpad-public"
that the data files have been uploaded to the google storage server.

Change-Id: I4d71a109fd53fae0c7c821bbac635f7d23c735dc
Reviewed-on: https://gerrit.chromium.org/gerrit/46631
Reviewed-by: Mike Truty <truty@chromium.org>
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>