firmware_TouchpadMTB: strip off the header of a gesture data file
In the header of a gesture data file, there is information about
the manufacturer and the touch device product. Sometimes, this
information may be sensitive and it is desirable to strip them off.

BUG=chromium:224228
TEST=perform the firmware tests.
$ cd /usr/local/autotest/tests/firmware_TouchpadMTB
$ DISPLAY=:0 python main.py

After the tests have been finished, check that the header of
gesture data files have been stripped off.

Change-Id: Id87966f3639da86f2a9c2fd8b8ba908ea039346f
Reviewed-on: https://gerrit.chromium.org/gerrit/46630
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>