Fix bug where full_release_test does not work well on trybots.
There's a couple things here. First we pass in a custom URL to find payloads
but we don't use it elsewhere. This never quite worked before (I was just lucky
because other canaries had run with that same version so source_image_uri was
always found from them). Instead I've fixed the logic to find archive_url for
the target image once (when we instantiate the test config generator) rather
than each time we find the payload uri.

Second, I've loosened the logic about exact matches for target releases as the
version may have an attempt # that is not passed in by the caller. If the
tested_release is a prefix of the version, it matches. I've also applied
this to the test because otherwise this test won't work for non-release images.
Non-release images use suffix's and timestamps that are pretty much impossible
to find out from a payload.

Finally, I've gotten rid of 3 unused variables from the test to cleanup and
make this logic simpler. We don't use the branches anymore in the test or
board. We just use the versions to verify that the test installed the right
image. Nothing more. The use for them was removed when I started inferring
the URL of the payload from the url's rather than re-constructing them.

BUG=chromium:224674
TEST=Ran it locally with nmo, npo, and specific on trybots and non-trybots.

Change-Id: If4c903de76d16ccb0d798bd09031aa3ac9f107a0
Reviewed-on: https://gerrit.chromium.org/gerrit/46808
Commit-Queue: Chris Sosa <sosa@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>