Fine-tune one test runner to work for one test instead of one suite.
LTP uses CMDFILES to designate suites.  The CMDFILE lists individual
tests (each with a unique tag).  There is no command line option on
runltp for an individual test.  This writes a cros_suite file with
the names supplied so that any one (set) of tests may be run.

This is especially helpful when debugging problems.

BUG=chromium:219105
TEST=./run_remote_tests.sh --remote=10.42.0.10 --use_emerged --board=link -a \
    "test_names=cros_unittest_sample"
    client/site_tests/kernel_LTP/control.select_tests
TEST=/run_remote_tests.sh --remote=10.42.0.10 --use_emerged --board=link -a \
    "test_names=cros_unittest_sample,cros_sample_getcwd02"
    client/site_tests/kernel_LTP/control.select_tests

Change-Id: Ib507ed1c7281aae850232bc1eedaf404667cfc39
Reviewed-on: https://gerrit.chromium.org/gerrit/46791
Commit-Queue: Mike Truty <truty@chromium.org>
Reviewed-by: Mike Truty <truty@chromium.org>
Tested-by: Mike Truty <truty@chromium.org>