Add temporary debugging logic around the call to "crossystem hwid".
The call to "crossystem hwid" is sometimes failing on certain
platforms.  To help in debugging the issue, we now catch the raised
exception and log the output of a second command before re-raising the
exception.

BUG=chromium:223728
TEST=Verified on my local device that if the original call (crossystem hwid)
fails, then the output of the second command is logged before the original
exception is re-raised.

Change-Id: I3184f013f7fbc2d304733552d1ee89adf653a6b9
Reviewed-on: https://gerrit.chromium.org/gerrit/46793
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>
Reviewed-by: Dennis Jeffrey <dennisjeffrey@chromium.org>