SyncControl: test eglGetSyncValuesCHROMIUM return value
If the X server returns and error for an DRI2GetMSC (e.g., "BadDrawable")
command, then eglGetSyncValuesCHROMIUM will return false, and the sync
values will be uninitialized.

This currently happens, for example, when attempting to get sync values
for a non-fullscreen X window, or when there are multiple active X windows
for the same X server (e.g., running synccontroltest on the same X server
as chrome).

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=The normal test should be unaffected:
     ./run_remote_tests.sh --use_emerged --board=daisy --remote=172.30.210.166 "graphics_SyncControlTest"
TEST=However, running manually on the device with the ui up should fail:
TEST=From ssh session run:
  # DISPLAY=:0 /usr/local/autotest/deps/glbench/synccontroltest
  This Should result in lots of messages like:
    Failure: First eglGetSyncValuesCHROMIUM returned false.

Change-Id: Ia15f5259b75654dba9b2d560bb0dd92476f43008
Reviewed-on: https://gerrit.chromium.org/gerrit/46865
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>