[Autotest] gs_offloader uses an rpc to query database.
gs_offloader currently creates a connection to the database
at import time and never forgets it. This means if the database
is migrated and gs_offloader isn't restarted it will continue to
ping the old database for the status of a job which doesn't exist.
This change goes through an rpc instead, forcing a new connection.

TEST=Forced a verify and checked True/False status
     of job by invoking is_special_task_complete directly.
BUG=chromium:225017

Change-Id: I5fefc0c086d8efb40028bafa7a2a961f4753d57a
Reviewed-on: https://gerrit.chromium.org/gerrit/46970
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>