autotest: Fix 095_ProfileGUID
This test has failed at least once in the lab because:

1) We configure the service (and set autoconnect to true)
2) We do some other work
3) We fail to connect the service because its already been
    autoconnected.

Fix this by not configuring autoconnect, and wait for our signal to
connect the service.  Since we're not testing autoconnect, this should
be just fine.

TEST=Test passes.  Unfortunately, this is fixing a race condition that I
only saw in logs.  The real test is just to see how this test does over
time in the lab.
BUG=chromium:225592

Change-Id: I95ec6cae919946b6f7c893cfdf5c53be6d668f02
Reviewed-on: https://gerrit.chromium.org/gerrit/47079
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>