autotest: Disable desktop_VideoSanity in BVT
Re land this change as the test is still failing, now on more boards
than just stout.

see gerrit change 46138:

This test fails with a kernel crash on Stout, which in turn causes
Autotest to fail the Canary build rather than ignoring the result
because of the experimental label.  There is a already a tracking high
priority bug to fix this error, and the TPMs are fine with releasing
canaries with this bug on Stout.

BUG=chromium:187993
TEST=none, this removes a test.

Change-Id: I0d4fc0d7989af384f4cd653b87def7bed1cc3bab
Reviewed-on: https://gerrit.chromium.org/gerrit/47242
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>