autotest: Fix remote_command to calculate pkill args correctly
Previously, when given a command 'ping 192.168.1.1', remote command
would hang indefinitely when the pkill failed to correctly kill the
ping command running in the background.  This is no longer the case.

Add a timeout to catch those cases where something odd goes wrong and
we have to abort anyway.  This is infinitely desirable to hanging
indefinitely.

TEST=Ran network_WiFiMatFunc/090_BgScanBackoff without it hanging.
     Backed out the changes to how we extract the basename, and noted
     that the test fails with an error about timing out.
BUG=None

Change-Id: I9a3c9acf2b0dcd24d7bc51117f1b4b9291495fe8
Reviewed-on: https://gerrit.chromium.org/gerrit/47255
Commit-Queue: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>