Touchpad FW Test: update gesture descriptions
This CL got reverted as it was suspected of causing a problem in the
automatic testing, but it turns out it wasn't this one.  This just adds
it back in.

Some of the gesture descriptions were a bit ambiguous apparently,
as some people were interpreting them differently.  This CL hopefully
clears them up and makes them a little clearer and more precise.

BUG=chromium:227021
TEST=manually ran

Change-Id: I1ae2a26226ab3eb4f71da5195abfa4ae9e7033fc
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47647
Reviewed-by: mukesh agrawal <quiche@chromium.org>