Adding a test for tap sensitivity to FW test
This adds a test to the Touch FW test that has the robot tap the screen
a bunch of times and test how many touches the sensor reported.  The
idea is that sometimes taps go unreported by the sensor, and we want to
make sure it's not the fault of the FW, since it might be incorrectly
filtering short taps as noise, or the hardware could be missing them
entirely.

The current test only has one validator, but we may expand upon that in
another CL if need be.

BUG=chrome-os-partner:17274
TEST=I installed the test on a Lumpy and ran it.

Change-Id: I98ab86f49dfe6b4b02a0f5af258eeb28d9f38e19
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47667
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>